macugEx/engine-server/engine-server/requirements.txt
flask==0.10.1
gevent==1.1.2
pyzmq==15.2.0
pymongo==3.2.2
paho-mqtt==1.2
apns-client==0.2.1
msgpack-python==0.4.7
redis==2.10.5
